move chdir to utils
